requirements.txt
torch>=1.1.0
tqdm>=4.31.1
pytorch_pretrained_bert==0.6.2
numpy>=1.15.4
python_Levenshtein>=0.12.0
prettytable